+++ server/src/modules/server/core.routes.ts
@@ -2,6 +2,7 @@ import config from "@/config";
 
 import { SResGetHealthCheck } from "../../../../shared/routes/core.routes";
 import packageJson from "../../../package.json";
+import { ensureInitialization } from "../../common/utils/mongodbUtils";
 import { Server } from "./server";
 
 export const coreRoutes = ({ server }: { server: Server }) => {
@@ -9,6 +10,7 @@ export const coreRoutes = ({ server }: { server: Server }) => {
     "/healthcheck",
     { schema: { response: { 200: SResGetHealthCheck } } },
     async (request, response) => {
+      await ensureInitialization();
       response.status(200).send({
         name: "[BAL] Apprentissage API",
         version: packageJson.version,