+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.45.3](https://github.com/mission-apprentissage/bal/compare/v1.45.2...v1.45.3) (2023-07-13)
+
+
+### Bug Fixes
+
+* mongodb connection ([#229](https://github.com/mission-apprentissage/bal/issues/229)) ([fbd1419](https://github.com/mission-apprentissage/bal/commit/fbd1419e775d2d3aa4b0ae93d851df7e1e56df75))
+
+## [1.45.2](https://github.com/mission-apprentissage/bal/compare/v1.45.1...v1.45.2) (2023-07-12)
+
+
+### Bug Fixes
+
+* migration status extension ([#228](https://github.com/mission-apprentissage/bal/issues/228)) ([180eec4](https://github.com/mission-apprentissage/bal/commit/180eec435e0ff27fce313c8b3a0ebedbd3c47560))
+
+## [1.45.1](https://github.com/mission-apprentissage/bal/compare/v1.45.0...v1.45.1) (2023-07-12)
+
+
+### Bug Fixes
+
+* CLI migrations scripts ([#227](https://github.com/mission-apprentissage/bal/issues/227)) ([8f5918e](https://github.com/mission-apprentissage/bal/commit/8f5918e09782efb0e3586097c658d6bae9aff7f0))
+
 # [1.45.0](https://github.com/mission-apprentissage/bal/compare/v1.44.0...v1.45.0) (2023-07-11)